fix(warehouse,reliefItem,Category): add caches and logs for all, fix some logic

+++ FloodRescue/FloodRescue.Services/Implements/Warehouse/WarehouseService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Confluent.Kafka;
 using FloodRescue.Repositories.Entites;
 using FloodRescue.Repositories.Interface;
 using FloodRescue.Services.DTO.Request.Warehouse;
@@ -23,7 +24,7 @@ namespace FloodRescue.Services.Implements.Warehouse
         private readonly IUnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
         private readonly ICacheService _cacheService;
-        private readonly ILogger<WarehouseService> _logger; 
+        private readonly ILogger<WarehouseService> _logger;
 
         public WarehouseService(IUnitOfWork unitOfWork, IMapper mapper, ICacheService cacheService, ILogger<WarehouseService> logger)
         {
@@ -47,35 +48,70 @@ namespace FloodRescue.Services.Implements.Warehouse
 
             CreateWarehouseResponseDTO responseDTO = _mapper.Map<CreateWarehouseResponseDTO>(warehouse);
 
-            await _cacheService.RemoveAsync(ALL_WAREHOUSES_KEY);    
+            await _cacheService.RemoveAsync(ALL_WAREHOUSES_KEY);
 
             return responseDTO;
         }
 
         public async Task<bool> DeleteWarehouseAsync(int warehouseId)
         {
+            _logger.LogInformation("Request to delete Warehouse ID: {WarehouseId}", warehouseId);
             WarehouseEntity? warehouse = await _unitOfWork.Warehouses.GetAsync(w => w.WarehouseID == warehouseId);
             int result = 0;
             if (warehouse != null)
             {
-                if (warehouse.IsDeleted == false)
-                {
-                    warehouse.IsDeleted = true;
-                    result = await _unitOfWork.SaveChangesAsync();
-                }
+                _logger.LogWarning("Delete failed. Warehouse ID: {WarehouseId} not found.", warehouseId);
+                return false;
             }
-            return result > 0;
+            if (warehouse.IsDeleted)
+            {
+                _logger.LogInformation("Warehouse ID: {WarehouseId} was already deleted. No changes made.", warehouseId);
+                return false;
+            }
+            warehouse.IsDeleted = true;
+            result = await _unitOfWork.SaveChangesAsync();
+            if (result > 0)
+            {
+                _logger.LogInformation("Successfully soft-deleted Warehouse ID: {WarehouseId} in database.", warehouseId);
+                await Task.WhenAll(
+                    _cacheService.RemoveAsync(ALL_WAREHOUSES_KEY),
+                    _cacheService.RemoveAsync($"{WAREHOUSE_KEY_PREFIX}{warehouseId}")
+                );
+                _logger.LogInformation("Cleared cache for Warehouse ID: {WarehouseId} and List.", warehouseId);
+                return true;
+            }
+            else
+            {
+                _logger.LogError("Unexpected error: Changes were not saved for Warehouse ID: {WarehouseId}", warehouseId);
+                return false;
+            }
         }
         public async Task<ShowWareHouseResponseDTO?> SearchWarehouseAsync(int id)
         {
-            ShowWareHouseResponseDTO? responseDTO = null;
+            // 1. kiểm tra cache trước tránh gọi db không cần thiết
+            _logger.LogInformation("Searching for Warehouse with ID: {WarehouseID}", id);
+            var cache = await _cacheService.GetAsync<ShowWareHouseResponseDTO>($"{WAREHOUSE_KEY_PREFIX}{id}");
+            if (cache != null)
+            {
+                _logger.LogInformation("Found Warehouse in cache: {WarehouseID}", id);
+                return cache;
+            }
+            _logger.LogInformation("Cache miss. Searching DB for Warehouse with ID: {WarehouseID}", id);
+            // 2. nếu không có trong cache thì gọi db
             WarehouseEntity? warehouse = await _unitOfWork.Warehouses.GetAsync(
                 w => w.WarehouseID == id && !w.IsDeleted,
                 w => w.Manager!
             );
+
+            ShowWareHouseResponseDTO? responseDTO = null;
+
             if (warehouse != null)
             {
                 responseDTO = _mapper.Map<ShowWareHouseResponseDTO>(warehouse);
+                // 3. Lưu vào Cache để lần sau không phải gọi DB nữa
+                // và set thời gian hết hạn (expiry) tùy theo nghiệp vụ
+                await _cacheService.SetAsync($"{WAREHOUSE_KEY_PREFIX}{id}", responseDTO, TimeSpan.FromMinutes(5));
+                _logger.LogInformation("Added Warehouse to cache: {WarehouseID}", id);
             }
             return responseDTO;
         }
@@ -86,7 +122,7 @@ namespace FloodRescue.Services.Implements.Warehouse
 
             var cached = await _cacheService.GetAsync<List<ShowWareHouseResponseDTO>>(ALL_WAREHOUSES_KEY);
 
-            if(cached != null)
+            if (cached != null)
             {
 
                 return cached;
@@ -110,14 +146,18 @@ namespace FloodRescue.Services.Implements.Warehouse
 
         public async Task<UpdateWarehouseResponseDTO> UpdateWarehouseAsync(int id, UpdateWarehouseRequestDTO request)
         {
+            _logger.LogInformation("Request to update Warehouse ID: {WarehouseId}", id);
             WarehouseEntity? _warehouse = await _unitOfWork.Warehouses.GetAsync(w => w.WarehouseID == id);
 
             if (_warehouse == null)
             {
+                _logger.LogWarning("Update failed. Warehouse ID: {WarehouseId} not found.", id);
                 return null;
             }
+            string oldName = _warehouse.Name;
 
-            WarehouseEntity newWarehouse = _mapper.Map<WarehouseEntity>(request);
+            _mapper.Map(request, _warehouse);
+            //WarehouseEntity newWarehouse = _mapper.Map<WarehouseEntity>(request);
             //_warehouse.Name = request.Name;
             //_warehouse.Address = request.Address;
             //_warehouse.LocationLong = request.LocationLong;
@@ -126,14 +166,22 @@ namespace FloodRescue.Services.Implements.Warehouse
 
             int result = await _unitOfWork.SaveChangesAsync();
 
+            _logger.LogInformation("Successfully updated Warehouse ID: {Id} changed name from '{OldName}' to '{NewName}'",
+            id, oldName, request.Name);
+
             if (result > 0)
             {
-                await _cacheService.RemoveAsync(ALL_WAREHOUSES_KEY);
-                await _cacheService.RemoveAsync($"{WAREHOUSE_KEY_PREFIX}{id}"); 
-
-                return _mapper.Map<UpdateWarehouseResponseDTO>(newWarehouse);
+                _logger.LogInformation("Successfully updated Warehouse ID: {WarehouseId} in database.", id);
+                await Task.WhenAll(
+                        _cacheService.RemoveAsync(ALL_WAREHOUSES_KEY),
+                        _cacheService.RemoveAsync($"{WAREHOUSE_KEY_PREFIX}{id}")
+                );
+                _logger.LogInformation("Cleared cache for Warehouse ID: {WarehouseId} and List.", id);
+                return _mapper.Map<UpdateWarehouseResponseDTO>(_warehouse);
             }
+            _logger.LogInformation("No changes detected for Warehouse ID: {WarehouseId}.", id);
             return null;
+
         }
     }
 }
